temp_repos/Hackathon/requirements.txt
streamlit
streamlit-folium
folium
twilio
python-dotenv
requests
Pillow
groq
camel-ai
SpeechRecognition
pyttsx3
geopy
easyocr
